apps/agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kimeboard-agent"
version = "0.1.0"
description = "Kimeboard agent service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "httpx>=0.28.1",
  "pydantic>=2.11.7",
  "pydantic-settings>=2.10.1",
  "structlog>=25.4.0",
  "tenacity>=9.1.2",
  "python-dateutil>=2.9.0.post0",
  "google-auth>=2.40.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
  "pytest-asyncio>=1.1.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
asyncio_mode = "auto"